fix(tree): colour
